make sure the degree and derivs are integers in attributes

+++ src/BSpline_export.cpp
@@ -52,7 +52,7 @@ Rcpp::NumericMatrix rcpp_bSpline_basis(
         R_NilValue, splines2::char_seq_len(out.ncol())
         );
     out.attr("x") = splines2::arma2rvec(x);
-    out.attr("degree") = bs_obj.get_degree();
+    out.attr("degree") = static_cast<int>(bs_obj.get_degree());
     out.attr("knots") = splines2::arma2rvec(bs_obj.get_internal_knots());
     out.attr("Boundary.knots") =
         splines2::arma2rvec(bs_obj.get_boundary_knots());
@@ -94,8 +94,8 @@ Rcpp::NumericMatrix rcpp_bSpline_derivative(
         R_NilValue, splines2::char_seq_len(out.ncol())
         );
     out.attr("x") = splines2::arma2rvec(x);
-    out.attr("derivs") = derivs;
-    out.attr("degree") = bs_obj.get_degree();
+    out.attr("derivs") = static_cast<int>(derivs);
+    out.attr("degree") = static_cast<int>(bs_obj.get_degree());
     out.attr("knots") = splines2::arma2rvec(bs_obj.get_internal_knots());
     out.attr("Boundary.knots") =
         splines2::arma2rvec(bs_obj.get_boundary_knots());
@@ -136,7 +136,7 @@ Rcpp::NumericMatrix rcpp_bSpline_integral(
         R_NilValue, splines2::char_seq_len(out.ncol())
         );
     out.attr("x") = splines2::arma2rvec(x);
-    out.attr("degree") = bs_obj.get_degree();
+    out.attr("degree") = static_cast<int>(bs_obj.get_degree());
     out.attr("knots") = splines2::arma2rvec(bs_obj.get_internal_knots());
     out.attr("Boundary.knots") =
         splines2::arma2rvec(bs_obj.get_boundary_knots());
